Reset Cinemachine camera, working on Player rotation
Reset the Cinemachine camera, it can zoom in and out, still need to make it automatically zoom in when close to an object. The camera can also rotate around the player without the player automatically looking where the camera is looking. Player currently doesn't face where it's moving. Working on that next.